+++ src/ddareungi_rl/algorithms/double_dqn.py
@@ -66,7 +66,7 @@ def train_double_dqn(
         episode_reward = 0.0
         episode_unmet = 0
         episode_rejected_returns = 0
-        episode_movement_cost = 0
+        episode_movement_cost = 0.0
         episode_losses = []
 
         while not done:
@@ -80,7 +80,7 @@ def train_double_dqn(
             episode_reward += reward
             episode_unmet += int(info["unmet_demand"])
             episode_rejected_returns += int(info["rejected_returns"])
-            episode_movement_cost += int(info["movement_cost"])
+            episode_movement_cost += float(info["movement_cost"])
 
             if len(replay) >= config.min_replay:
                 episode_losses.append(
